New PHP Redirects
After redirects are checked, published and the cache is cleared, they need to be marked as fixed in GSC (Google Search Console).

Many redirects in the original list have not been included as they look like on page errors.

If you can provide me with the 404 source list (available in GSC) I can check and fix these pages.

Other redirects not included are the case-sensitive ones. Permalinks shouldn't be case-sensitive. These 404 pages will be fixed automatically when the permalinks are no longer case sensitive.